requirement.txt
greenlet==1.1.1
numpy==1.21.2
pandas==1.3.2
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
